requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mongoengine==0.18.2
pymongo==3.8.0
six==1.12.0
Werkzeug==0.15.5
python-dotenv==0.10.3
